update kaitai definitions
