+++ .gitignore
@@ -14,7 +14,7 @@ MANIFEST
 .tox*/
 build/
 dist/
-mitmproxy/contrib/kaitaistruct/*.ksy
+mitmproxy/contrib/kaitaistruct/**/*.ksy
 .pytest_cache
 __pycache__
 .hypothesis/
