+++ mitmproxy/contrib/kaitaistruct/google_protobuf.py
@@ -1,13 +1,14 @@
 # This is a generated file! Please edit source .ksy file and use kaitai-struct-compiler to rebuild
 
-
+import kaitaistruct
 from kaitaistruct import KaitaiStruct, KaitaiStream, BytesIO
 from enum import Enum
 
 
-# manually removed version check, see https://github.com/mitmproxy/mitmproxy/issues/5401
+if getattr(kaitaistruct, 'API_VERSION', (0, 9)) < (0, 9):
+    raise Exception("Incompatible Kaitai Struct Python API: 0.9 or later is required, but you have %s" % (kaitaistruct.__version__))
 
-from .vlq_base128_le import VlqBase128Le
+from . import vlq_base128_le
 class GoogleProtobuf(KaitaiStruct):
     """Google Protocol Buffers (AKA protobuf) is a popular data
     serialization scheme used for communication protocols, data storage,
@@ -48,8 +49,10 @@ class GoogleProtobuf(KaitaiStruct):
 
     def _read(self):
         self.pairs = []
+        i = 0
         while not self._io.is_eof():
-            self.pairs.append(self._root.Pair(self._io, self, self._root))
+            self.pairs.append(GoogleProtobuf.Pair(self._io, self, self._root))
+            i += 1
 
 
     class Pair(KaitaiStruct):
@@ -69,15 +72,15 @@ class GoogleProtobuf(KaitaiStruct):
             self._read()
 
         def _read(self):
-            self.key = VlqBase128Le(self._io)
+            self.key = vlq_base128_le.VlqBase128Le(self._io)
             _on = self.wire_type
-            if _on == self._root.Pair.WireTypes.varint:
-                self.value = VlqBase128Le(self._io)
-            elif _on == self._root.Pair.WireTypes.len_delimited:
-                self.value = self._root.DelimitedBytes(self._io, self, self._root)
-            elif _on == self._root.Pair.WireTypes.bit_64:
+            if _on == GoogleProtobuf.Pair.WireTypes.varint:
+                self.value = vlq_base128_le.VlqBase128Le(self._io)
+            elif _on == GoogleProtobuf.Pair.WireTypes.len_delimited:
+                self.value = GoogleProtobuf.DelimitedBytes(self._io, self, self._root)
+            elif _on == GoogleProtobuf.Pair.WireTypes.bit_64:
                 self.value = self._io.read_u8le()
-            elif _on == self._root.Pair.WireTypes.bit_32:
+            elif _on == GoogleProtobuf.Pair.WireTypes.bit_32:
                 self.value = self._io.read_u4le()
 
         @property
@@ -91,10 +94,10 @@ class GoogleProtobuf(KaitaiStruct):
             arbitrary bytes from UTF-8 encoded strings, etc.
             """
             if hasattr(self, '_m_wire_type'):
-                return self._m_wire_type if hasattr(self, '_m_wire_type') else None
+                return self._m_wire_type
 
-            self._m_wire_type = self._root.Pair.WireTypes((self.key.value & 7))
-            return self._m_wire_type if hasattr(self, '_m_wire_type') else None
+            self._m_wire_type = KaitaiStream.resolve_enum(GoogleProtobuf.Pair.WireTypes, (self.key.value & 7))
+            return getattr(self, '_m_wire_type', None)
 
         @property
         def field_tag(self):
@@ -102,10 +105,10 @@ class GoogleProtobuf(KaitaiStruct):
             field name in a `.proto` file by this field tag.
             """
             if hasattr(self, '_m_field_tag'):
-                return self._m_field_tag if hasattr(self, '_m_field_tag') else None
+                return self._m_field_tag
 
             self._m_field_tag = (self.key.value >> 3)
-            return self._m_field_tag if hasattr(self, '_m_field_tag') else None
+            return getattr(self, '_m_field_tag', None)
 
 
     class DelimitedBytes(KaitaiStruct):
@@ -116,7 +119,7 @@ class GoogleProtobuf(KaitaiStruct):
             self._read()
 
         def _read(self):
-            self.len = VlqBase128Le(self._io)
+            self.len = vlq_base128_le.VlqBase128Le(self._io)
             self.body = self._io.read_bytes(self.len.value)
