+++ mitmproxy/contrib/kaitaistruct/ico.py
@@ -1,10 +1,11 @@
 # This is a generated file! Please edit source .ksy file and use kaitai-struct-compiler to rebuild
 
+import kaitaistruct
 from kaitaistruct import KaitaiStruct, KaitaiStream, BytesIO
-import struct
 
 
-# manually removed version check, see https://github.com/mitmproxy/mitmproxy/issues/5401
+if getattr(kaitaistruct, 'API_VERSION', (0, 9)) < (0, 9):
+    raise Exception("Incompatible Kaitai Struct Python API: 0.9 or later is required, but you have %s" % (kaitaistruct.__version__))
 
 class Ico(KaitaiStruct):
     """Microsoft Windows uses specific file format to store applications
@@ -13,7 +14,7 @@ class Ico(KaitaiStruct):
     contained inside).
     
     .. seealso::
-       Source - https://msdn.microsoft.com/en-us/library/ms997538.aspx
+       Source - https://docs.microsoft.com/en-us/previous-versions/ms997538(v=msdn.10)
     """
     def __init__(self, _io, _parent=None, _root=None):
         self._io = _io
@@ -22,11 +23,13 @@ class Ico(KaitaiStruct):
         self._read()
 
     def _read(self):
-        self.magic = self._io.ensure_fixed_contents(struct.pack('4b', 0, 0, 1, 0))
+        self.magic = self._io.read_bytes(4)
+        if not self.magic == b"\x00\x00\x01\x00":
+            raise kaitaistruct.ValidationNotEqualError(b"\x00\x00\x01\x00", self.magic, self._io, u"/seq/0")
         self.num_images = self._io.read_u2le()
-        self.images = [None] * (self.num_images)
+        self.images = []
         for i in range(self.num_images):
-            self.images[i] = self._root.IconDirEntry(self._io, self, self._root)
+            self.images.append(Ico.IconDirEntry(self._io, self, self._root))
 
 
     class IconDirEntry(KaitaiStruct):
@@ -40,7 +43,9 @@ class Ico(KaitaiStruct):
             self.width = self._io.read_u1()
             self.height = self._io.read_u1()
             self.num_colors = self._io.read_u1()
-            self.reserved = self._io.ensure_fixed_contents(struct.pack('1b', 0))
+            self.reserved = self._io.read_bytes(1)
+            if not self.reserved == b"\x00":
+                raise kaitaistruct.ValidationNotEqualError(b"\x00", self.reserved, self._io, u"/types/icon_dir_entry/seq/3")
             self.num_planes = self._io.read_u2le()
             self.bpp = self._io.read_u2le()
             self.len_img = self._io.read_u4le()
@@ -53,13 +58,13 @@ class Ico(KaitaiStruct):
             relevant parser, if needed to parse image data further.
             """
             if hasattr(self, '_m_img'):
-                return self._m_img if hasattr(self, '_m_img') else None
+                return self._m_img
 
             _pos = self._io.pos()
             self._io.seek(self.ofs_img)
             self._m_img = self._io.read_bytes(self.len_img)
             self._io.seek(_pos)
-            return self._m_img if hasattr(self, '_m_img') else None
+            return getattr(self, '_m_img', None)
 
         @property
         def png_header(self):
@@ -67,22 +72,22 @@ class Ico(KaitaiStruct):
             embedded PNG file.
             """
             if hasattr(self, '_m_png_header'):
-                return self._m_png_header if hasattr(self, '_m_png_header') else None
+                return self._m_png_header
 
             _pos = self._io.pos()
             self._io.seek(self.ofs_img)
             self._m_png_header = self._io.read_bytes(8)
             self._io.seek(_pos)
-            return self._m_png_header if hasattr(self, '_m_png_header') else None
+            return getattr(self, '_m_png_header', None)
 
         @property
         def is_png(self):
             """True if this image is in PNG format."""
             if hasattr(self, '_m_is_png'):
-                return self._m_is_png if hasattr(self, '_m_is_png') else None
+                return self._m_is_png
 
-            self._m_is_png = self.png_header == struct.pack('8b', -119, 80, 78, 71, 13, 10, 26, 10)
-            return self._m_is_png if hasattr(self, '_m_is_png') else None
+            self._m_is_png = self.png_header == b"\x89\x50\x4E\x47\x0D\x0A\x1A\x0A"
+            return getattr(self, '_m_is_png', None)
