+++ mitmproxy/contrib/kaitaistruct/tls_client_hello.ksy
@@ -1,39 +1,37 @@
 meta:
   id: tls_client_hello
+  xref:
+    rfc: 5246 # TLS 1.2
+    wikidata: Q206494 # TLS
+  license: MIT
   endian: be
 
 seq:
   - id: version
     type: version
-  
+
   - id: random
     type: random
-    
+
   - id: session_id
     type: session_id
-    
+
   - id: cipher_suites
     type: cipher_suites
-  
+
   - id: compression_methods
     type: compression_methods
-  
-  - id: extensions
-    size: 0
-    repeat: expr
-    repeat-expr: 0
-    if: _io.eof == true
 
   - id: extensions
     type: extensions
     if: _io.eof == false
-    
+
 types:
   version:
     seq:
       - id: major
         type: u1
-        
+
       - id: minor
         type: u1
 
@@ -41,7 +39,7 @@ types:
     seq:
       - id: gmt_unix_time
         type: u4
-      
+
       - id: random
         size: 28
 
@@ -49,15 +47,15 @@ types:
     seq:
       - id: len
         type: u1
-        
+
       - id: sid
         size: len
-  
+
   cipher_suites:
     seq:
       - id: len
         type: u2
-      
+
       - id: cipher_suites
         type: u2
         repeat: expr
@@ -67,7 +65,7 @@ types:
     seq:
       - id: len
         type: u1
-      
+
       - id: compression_methods
         size: len
 
@@ -79,23 +77,23 @@ types:
       - id: extensions
         type: extension
         repeat: eos
-  
+
   extension:
     seq:
       - id: type
         type: u2
-      
+
       - id: len
         type: u2
-        
+
       - id: body
         size: len
-        type: 
+        type:
           switch-on: type
           cases:
             0: sni
             16: alpn
-  
+
   sni:
     seq:
       - id: list_length
@@ -104,12 +102,12 @@ types:
       - id: server_names
         type: server_name
         repeat: eos
-  
+
   server_name:
     seq:
       - id: name_type
         type: u1
-  
+
       - id: length
         type: u2
 
@@ -129,6 +127,6 @@ types:
     seq:
       - id: strlen
         type: u1
-        
+
       - id: name
         size: strlen
