+++ mitmproxy/contrib/kaitaistruct/vlq_base128_le.py
@@ -1,13 +1,15 @@
 # This is a generated file! Please edit source .ksy file and use kaitai-struct-compiler to rebuild
 
+import kaitaistruct
 from kaitaistruct import KaitaiStruct, KaitaiStream, BytesIO
 
 
-# manually removed version check, see https://github.com/mitmproxy/mitmproxy/issues/5401
+if getattr(kaitaistruct, 'API_VERSION', (0, 9)) < (0, 9):
+    raise Exception("Incompatible Kaitai Struct Python API: 0.9 or later is required, but you have %s" % (kaitaistruct.__version__))
 
 class VlqBase128Le(KaitaiStruct):
-    """A variable-length unsigned integer using base128 encoding. 1-byte groups
-    consists of 1-bit flag of continuation and 7-bit value, and are ordered
+    """A variable-length unsigned/signed integer using base128 encoding. 1-byte groups
+    consist of 1-bit flag of continuation and 7-bit value chunk, and are ordered
     "least significant group first", i.e. in "little-endian" manner.
     
     This particular encoding is specified and used in:
@@ -17,10 +19,10 @@ class VlqBase128Le(KaitaiStruct):
     * Google Protocol Buffers, where it's called "Base 128 Varints".
       https://developers.google.com/protocol-buffers/docs/encoding?csw=1#varints
     * Apache Lucene, where it's called "VInt"
-      http://lucene.apache.org/core/3_5_0/fileformats.html#VInt
+      https://lucene.apache.org/core/3_5_0/fileformats.html#VInt
     * Apache Avro uses this as a basis for integer encoding, adding ZigZag on
       top of it for signed ints
-      http://avro.apache.org/docs/current/spec.html#binary_encode_primitive
+      https://avro.apache.org/docs/current/spec.html#binary_encode_primitive
     
     More information on this encoding is available at https://en.wikipedia.org/wiki/LEB128
     
@@ -34,15 +36,16 @@ class VlqBase128Le(KaitaiStruct):
 
     def _read(self):
         self.groups = []
+        i = 0
         while True:
-            _ = self._root.Group(self._io, self, self._root)
+            _ = VlqBase128Le.Group(self._io, self, self._root)
             self.groups.append(_)
             if not (_.has_next):
                 break
+            i += 1
 
     class Group(KaitaiStruct):
-        """One byte group, clearly divided into 7-bit "value" and 1-bit "has continuation
-        in the next byte" flag.
+        """One byte group, clearly divided into 7-bit "value" chunk and 1-bit "continuation" flag.
         """
         def __init__(self, _io, _parent=None, _root=None):
             self._io = _io
@@ -57,36 +60,56 @@ class VlqBase128Le(KaitaiStruct):
         def has_next(self):
             """If true, then we have more bytes to read."""
             if hasattr(self, '_m_has_next'):
-                return self._m_has_next if hasattr(self, '_m_has_next') else None
+                return self._m_has_next
 
             self._m_has_next = (self.b & 128) != 0
-            return self._m_has_next if hasattr(self, '_m_has_next') else None
+            return getattr(self, '_m_has_next', None)
 
         @property
         def value(self):
-            """The 7-bit (base128) numeric value of this group."""
+            """The 7-bit (base128) numeric value chunk of this group."""
             if hasattr(self, '_m_value'):
-                return self._m_value if hasattr(self, '_m_value') else None
+                return self._m_value
 
             self._m_value = (self.b & 127)
-            return self._m_value if hasattr(self, '_m_value') else None
+            return getattr(self, '_m_value', None)
 
 
     @property
     def len(self):
         if hasattr(self, '_m_len'):
-            return self._m_len if hasattr(self, '_m_len') else None
+            return self._m_len
 
         self._m_len = len(self.groups)
-        return self._m_len if hasattr(self, '_m_len') else None
+        return getattr(self, '_m_len', None)
 
     @property
     def value(self):
-        """Resulting value as normal integer."""
+        """Resulting unsigned value as normal integer."""
         if hasattr(self, '_m_value'):
-            return self._m_value if hasattr(self, '_m_value') else None
+            return self._m_value
 
         self._m_value = (((((((self.groups[0].value + ((self.groups[1].value << 7) if self.len >= 2 else 0)) + ((self.groups[2].value << 14) if self.len >= 3 else 0)) + ((self.groups[3].value << 21) if self.len >= 4 else 0)) + ((self.groups[4].value << 28) if self.len >= 5 else 0)) + ((self.groups[5].value << 35) if self.len >= 6 else 0)) + ((self.groups[6].value << 42) if self.len >= 7 else 0)) + ((self.groups[7].value << 49) if self.len >= 8 else 0))
-        return self._m_value if hasattr(self, '_m_value') else None
+        return getattr(self, '_m_value', None)
+
+    @property
+    def sign_bit(self):
+        if hasattr(self, '_m_sign_bit'):
+            return self._m_sign_bit
+
+        self._m_sign_bit = (1 << ((7 * self.len) - 1))
+        return getattr(self, '_m_sign_bit', None)
+
+    @property
+    def value_signed(self):
+        """
+        .. seealso::
+           Source - https://graphics.stanford.edu/~seander/bithacks.html#VariableSignExtend
+        """
+        if hasattr(self, '_m_value_signed'):
+            return self._m_value_signed
+
+        self._m_value_signed = ((self.value ^ self.sign_bit) - self.sign_bit)
+        return getattr(self, '_m_value_signed', None)
